dont delete last old backups! THX to Marco Lenhardt!

+++ backup-docker-mysql.sh
@@ -59,6 +59,9 @@ for i in $CONTAINER; do
 done
 echo -e "\n$TIMESTAMP Backup for Databases completed\n" 
 
-find $BACKUPDIR -name "*.gz" -daystart -mtime +$DAYS -delete
-
-
+# dont delete last old backups!
+OLD_BACKUPS=$(ls -1 $BACKUPDIR/*.gz |wc -l)
+if [ $OLD_BACKUPS -gt 3 ]
+then
+	find $BACKUPDIR -name "*.gz" -daystart -mtime +$DAYS -delete
+fi
